IDP-1971 Initializing refactor delete
