+++ src/main/java/it/gov/pagopa/payment/controller/payment/CommonPaymentControllerImpl.java
@@ -3,8 +3,10 @@ package it.gov.pagopa.payment.controller.payment;
 import it.gov.pagopa.common.performancelogger.PerformanceLog;
 import it.gov.pagopa.payment.dto.common.BaseTransactionResponseDTO;
 import it.gov.pagopa.payment.dto.qrcode.TransactionCreationRequest;
+import it.gov.pagopa.payment.service.payment.common.CommonCancelServiceImpl;
 import it.gov.pagopa.payment.service.payment.common.CommonCreationServiceImpl;
 import it.gov.pagopa.payment.service.performancelogger.BaseTransactionResponseDTOPerfLoggerPayloadBuilder;
+import it.gov.pagopa.payment.service.performancelogger.TransactionResponsePerfLoggerPayloadBuilder;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.web.bind.annotation.RestController;
@@ -12,9 +14,12 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class CommonPaymentControllerImpl implements CommonPaymentController {
     private final CommonCreationServiceImpl commonCreationService;
+    private final CommonCancelServiceImpl commonCancelService;
 
-    public CommonPaymentControllerImpl( @Qualifier("CommonCreate") CommonCreationServiceImpl commonCreationService) {
+    public CommonPaymentControllerImpl(@Qualifier("CommonCreate") CommonCreationServiceImpl commonCreationService,
+                                       @Qualifier("CommonCancel")CommonCancelServiceImpl commonCancelService) {
         this.commonCreationService = commonCreationService;
+        this.commonCancelService = commonCancelService;
     }
 
     @Override
@@ -28,5 +33,18 @@ public class CommonPaymentControllerImpl implements CommonPaymentController {
             String idTrxIssuer) {
         log.info("[CREATE_TRANSACTION] The merchant {} through acquirer {} is creating a transaction", merchantId, acquirerId);
         return commonCreationService.createTransaction(trxCreationRequest,null,merchantId,acquirerId,idTrxIssuer);
+    }
+
+    @Override
+    @PerformanceLog(
+            value = "CANCEL_TRANSACTION",
+            payloadBuilderBeanClass = TransactionResponsePerfLoggerPayloadBuilder.class)
+    public void cancelTransaction(String trxId, String merchantId, String acquirerId) {
+        log.info(
+                "[CANCEL_TRANSACTION] The merchant {} through acquirer {} is cancelling the transaction {}",
+                merchantId,
+                acquirerId,
+                trxId);
+        commonCancelService.cancelTransaction(trxId, merchantId, acquirerId);
     }
 }
