+++ src/test/java/it/gov/pagopa/payment/controller/payment/CommonPaymentControllerTest.java
@@ -6,6 +6,7 @@ import it.gov.pagopa.common.web.dto.ErrorDTO;
 import it.gov.pagopa.common.web.exception.ValidationExceptionHandler;
 import it.gov.pagopa.payment.dto.common.BaseTransactionResponseDTO;
 import it.gov.pagopa.payment.dto.qrcode.TransactionCreationRequest;
+import it.gov.pagopa.payment.service.payment.common.CommonCancelServiceImpl;
 import it.gov.pagopa.payment.service.payment.common.CommonCreationServiceImpl;
 import it.gov.pagopa.payment.test.fakers.BaseTransactionResponseFaker;
 import it.gov.pagopa.payment.test.fakers.TransactionCreationRequestFaker;
@@ -25,7 +26,9 @@ import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import java.util.Arrays;
 import java.util.List;
 
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @WebMvcTest(CommonPaymentControllerImpl.class)
@@ -34,6 +37,10 @@ class CommonPaymentControllerTest {
     @MockBean
     @Qualifier("CommonCreate")
     private CommonCreationServiceImpl commonCreationServiceMock;
+
+    @MockBean
+    @Qualifier("CommonCancel")
+    private CommonCancelServiceImpl commonCancelServiceMock;
     @Autowired
     private MockMvc mockMvc;
 
@@ -88,6 +95,25 @@ class CommonPaymentControllerTest {
         Assertions.assertNotNull(resultResponse);
         Assertions.assertEquals(response,resultResponse);
         Mockito.verify(commonCreationServiceMock).createTransaction(body,null,MERCHANT_ID,ACQUIRER_ID,ID_TRX_ISSUER);
+    }
+
+    @Test
+    void cancelTransaction_testMandatoryHeaders() throws Exception {
+
+        TransactionCreationRequest body = TransactionCreationRequestFaker.mockInstance(1);
+
+        MvcResult result = mockMvc.perform(
+                        delete("/idpay/payment/merchant/{transactionId}",
+                                "TRXID")
+                                .contentType(MediaType.APPLICATION_JSON))
+                .andExpect(status().isBadRequest())
+                .andReturn();
+
+        assertNotNull(result.getResponse().getContentAsString());
+
+        String actual = "{\"code\":\"INVALID_REQUEST\",\"message\":\"Required request header "
+                + "'x-merchant-id' for method parameter type String is not present\"}";
+        assertEquals(actual, result.getResponse().getContentAsString());
     }
 
 }
